+++ test/integration/failover/test_failover.py
@@ -6,11 +6,10 @@ from integration.failover.utils import (
     get_stateboard_failover_info,
 )
 
-from integration.replicasets.conftest import default_project_with_instances
-from conftest import project_without_dependencies
 from utils import run_command_and_get_output
 
 
+# Fixtures
 @pytest.fixture(scope="function")
 def project_with_topology(cartridge_cmd, default_project_with_instances, tmpdir):
     project = default_project_with_instances.project
@@ -22,6 +21,7 @@ def project_with_topology(cartridge_cmd, default_project_with_instances, tmpdir)
     return project
 
 
+# Tests
 def test_default_app_stateboard_failover(cartridge_cmd, project_with_topology):
     project = project_with_topology
 
@@ -160,6 +160,3 @@ def test_invalid_stateboard_failover_opts(cartridge_cmd):
 
 def test_invalid_etcd2_failover_opts(cartridge_cmd):
     pass
-
-
-
